Fixed the image for the main github README, quick overview of structure:
weatherMapTest and WunderGroundTest2 both rely on a WeatherController which contains a model of the weather information. Currently the logic for a client side checking of this information is written to the code and needs to be forked out. The main model is a weatherMap object which controls the precipitation imagery which is converted into the 2D array which is used to map a Boundary to a weather pixel. The Boundary is key as it is mapped to the same 2D array, but relative to a GPS location on the real world and map. This can be seen in the URL and the convertToXY method in MAIN and other derived classes, most of the map collectiong logic can be seen there, this will be the basis of the backend server, DOUG, not an acronym but the nickname of the Douglas fir adorning the Cascadia flag, Doug. Since my map currently covers a large portion of Cascadia it seemed only fitting that DOUG be the one watching over its current weather.

The boundary checking which will detect weather, is still very clunky and uses an as of yet somewhat unreliable custom linkedlist to iterate through the points on the Boundary edge(s), eventually several bounds will need to be checked in order to calculate time of arrival ,ToA, and and possibly more info. The preceding mentioned main classes use a boundary checker which iterates over this set of linked lists with Points in them, each point has a location relative to the center of the Boundary ac calculated in gpstoXY, these points match up with the 2D weather array information and tell it where to check on the array for percipitation. The idea of quadrants was to make checking threaded and give some granularity to where the check would start, possibly relative to calculated movement of weather fronts therefore being able to optimize the checking to follow the movement. Coverage is then calculated at the edge of the boundary, this however suffers from rounding errors and some non matching reports as seen in the tests for weatherMap, where the FULL/HALF/QUARTER maps are used, the reported coverages aren't as expected currently.

So client side is still a mess, and backend is esentially imobilized for the time being...did I forget anything?